pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = {}

[project]
name = "environ"
version = "0.0.1" 
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "pandas",
    'scipy',
    "requests",
    "matplotlib",
    "seaborn",
    "pyarrow",
    "tqdm",
    "psycopg2-binary",
    "python-dotenv",
    "tenacity",
    "d3blocks",
    "openai",
    "dovenv",
    "openpyxl",
    "shap",
]

[project.optional-dependencies]
dev = ["pylint"]